se corrigieron warnigns de compilación. Se testeó InicializarMV, cargarPrograma, ejecutarPrograma,  iniciaVectorFUnciones, cantidadOperandos, obtenerTiposOperandos, cargarOperandos y ejecutarInstrucción

+++ MV.h
@@ -26,6 +26,6 @@ typedef struct {
 void inicializarMV(MV *mv);
 void cargarPrograma(MV *mv, char *nombreArchivo);
 void ejecutarPrograma(MV *mv);
-void obtenerValorOperando(MV *mv, uint32_t operando, int32_t *valor);
+void obtenerValorOperando(MV *mv, uint32_t operando, long int *valor);
 void guardarValorOperando(MV *mv, long int operando, long int valor);
 #endif
